Split math.kt into math.kt and dateCalculations.kt
