+++ core/common/test/LocalDateTest.kt
@@ -6,6 +6,7 @@
 package kotlinx.datetime.test
 
 import kotlinx.datetime.*
+import kotlinx.datetime.internal.*
 import kotlin.random.*
 import kotlin.test.*
