+++ core/js/src/LocalTime.kt
@@ -4,6 +4,7 @@
  */
 package kotlinx.datetime
 
+import kotlinx.datetime.internal.*
 import kotlinx.datetime.serializers.LocalTimeIso8601Serializer
 import kotlinx.serialization.Serializable
 import kotlinx.datetime.internal.JSJoda.LocalTime as jtLocalTime
