+++ core/jvm/src/LocalDate.kt
@@ -5,6 +5,9 @@
 @file:JvmName("LocalDateJvmKt")
 package kotlinx.datetime
 
+import kotlinx.datetime.internal.safeAdd
+import kotlinx.datetime.internal.safeMultiply
+import kotlinx.datetime.internal.*
 import kotlinx.datetime.serializers.LocalDateIso8601Serializer
 import kotlinx.serialization.Serializable
 import java.time.DateTimeException
