+++ core/jvm/src/LocalTime.kt
@@ -6,6 +6,7 @@
 
 package kotlinx.datetime
 
+import kotlinx.datetime.internal.*
 import kotlinx.datetime.serializers.LocalTimeIso8601Serializer
 import kotlinx.serialization.Serializable
 import java.time.DateTimeException
