+++ core/native/cinterop_actuals/TimeZoneNative.kt
@@ -6,6 +6,7 @@ package kotlinx.datetime
 
 import kotlinx.datetime.internal.*
 import kotlinx.cinterop.*
+import kotlinx.datetime.internal.*
 import platform.posix.free
 
 internal actual class RegionTimeZone(private val tzid: TZID, actual override val id: String): TimeZone() {
