+++ core/native/src/internal/dateCalculations.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2019-2022 JetBrains s.r.o. and contributors.
+ * Use of this source code is governed by the Apache 2.0 License that can be found in the LICENSE.txt file.
+ */
+
+package kotlinx.datetime.internal
+
+import kotlin.math.*
+
+/**
+ * All code below was taken from various places of https://github.com/ThreeTen/threetenbp with few changes
+ */
+
+/**
+ * The number of days in a 400 year cycle.
+ */
+internal const val DAYS_PER_CYCLE = 146097
+
+/**
+ * The number of days from year zero to year 1970.
+ * There are five 400 year cycles from year zero to 2000.
+ * There are 7 leap years from 1970 to 2000.
+ */
+internal const val DAYS_0000_TO_1970 = DAYS_PER_CYCLE * 5 - (30 * 365 + 7)
+
+// days in a 400 year cycle = 146097
+// days in a 10,000 year cycle = 146097 * 25
+// seconds per day = 86400
+internal const val SECONDS_PER_10000_YEARS = 146097L * 25L * 86400L
+
+internal const val SECONDS_0000_TO_1970 = (146097L * 5L - (30L * 365L + 7L)) * 86400L
+
+// org.threeten.bp.ZoneOffset#buildId
+internal fun zoneIdByOffset(totalSeconds: Int): String {
+    return if (totalSeconds == 0) {
+        "Z"
+    } else {
+        val absTotalSeconds: Int = abs(totalSeconds)
+        val buf = StringBuilder()
+        val absHours: Int = absTotalSeconds / SECONDS_PER_HOUR
+        val absMinutes: Int = absTotalSeconds / SECONDS_PER_MINUTE % MINUTES_PER_HOUR
+        buf.append(if (totalSeconds < 0) "-" else "+")
+            .append(if (absHours < 10) "0" else "").append(absHours)
+            .append(if (absMinutes < 10) ":0" else ":").append(absMinutes)
+        val absSeconds: Int = absTotalSeconds % SECONDS_PER_MINUTE
+        if (absSeconds != 0) {
+            buf.append(if (absSeconds < 10) ":0" else ":").append(absSeconds)
+        }
+        buf.toString()
+    }
+}
