+++ core/native/test/ThreeTenBpLocalDateTest.kt
@@ -9,6 +9,7 @@
 package kotlinx.datetime.test
 
 import kotlinx.datetime.*
+import kotlinx.datetime.internal.*
 import kotlin.test.*
 
 class ThreeTenBpLocalDateTest {
