+++ core/native/test/ThreeTenBpUtilTest.kt
@@ -8,7 +8,7 @@
 
 package kotlinx.datetime.test
 
-import kotlinx.datetime.*
+import kotlinx.datetime.internal.*
 import kotlin.test.*
 
 class ThreeTenBpUtilTest {
